Update README with Credly badges as images
This commit updates badges from Credly as images to the README file, displaying my achievements and competencies in various fields. The badges are linked to the original badges on the Credly website, and are displayed in the README file using HTML tags and image tags. This update provides a quick overview of my skills and qualifications to anyone viewing my profile on GitHub.